Dec 4, 2022, 1:56 PM

+++ Assets/Script/RubyController.cs
@@ -38,10 +38,10 @@ public class RubyController : MonoBehaviour
 
     public float timeInvincible = 2.0f;
     public float timeInvincible2 = 15.0f;
-    public float timeBoosting = 1;
+    public int timeBoosting = 1;
     bool isInvincible;
     float invincibleTimer;
-    float isBoostingTimer;
+    int isBoostingTimer;
 
     bool cakeYes;
